client/src/requirements.txt
pygame
Pillow
numpy
